Add files via upload
This project is developed using bootstrap, jQuery, Google scripts and Google sheets. The user will be asked to take an online quiz. On submit the quiz results will be submitted to a google sheet and graded using google script. An email with quiz results will be sent to user immediately. You will need to have a gmail account for running the google script. You will have to create a google sheet and write the google script in the corresponding script.